Extract SystemPropertiesModification from SystemPropertiesExtension
