+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/util/SystemPropertiesExtension.java
@@ -10,29 +10,19 @@
 
 package org.junit.jupiter.api.util;
 
-import static java.util.stream.Collectors.toSet;
 import static org.junit.jupiter.api.util.JupiterPropertyUtils.cloneWithoutDefaults;
-import static org.junit.platform.commons.support.AnnotationSupport.findRepeatableAnnotations;
 import static org.junit.platform.commons.support.AnnotationSupport.isAnnotated;
-import static org.junit.platform.commons.util.CollectionUtils.forEachInReverseOrder;
 
-import java.lang.reflect.AnnotatedElement;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
-import java.util.Set;
 
 import org.junit.jupiter.api.extension.AfterAllCallback;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionConfigurationException;
 import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.platform.commons.util.ToStringBuilder;
 
 /**
  * {@code Extension} which provides support for the following annotations.
@@ -60,7 +50,7 @@ final class SystemPropertiesExtension
 
 	private void applyForAllContexts(ExtensionContext context) {
 		var allContexts = findAllExtensionContexts(context);
-		var modification = DeferredPropertyModification.create(allContexts);
+		var modification = SystemPropertiesModification.create(allContexts);
 
 		// Please do not refactor out the common parts.
 		findFirstRestoreAnnotationContext(allContexts) //
@@ -85,7 +75,7 @@ final class SystemPropertiesExtension
 		getStore(context).put(createStoreKey(context, BackupType.COMPLETE), backup);
 	}
 
-	private void storePartialBackup(ExtensionContext context, DeferredPropertyModification backup) {
+	private void storePartialBackup(ExtensionContext context, SystemPropertiesModification backup) {
 		getStore(context).put(createStoreKey(context, BackupType.PARTIAL), backup);
 	}
 
@@ -114,9 +104,9 @@ final class SystemPropertiesExtension
 		return Optional.ofNullable(backup);
 	}
 
-	private Optional<DeferredPropertyModification> findPartialBackup(ExtensionContext context) {
+	private Optional<SystemPropertiesModification> findPartialBackup(ExtensionContext context) {
 		var key = createStoreKey(context, BackupType.PARTIAL);
-		var backup = getStore(context).get(key, DeferredPropertyModification.class);
+		var backup = getStore(context).get(key, SystemPropertiesModification.class);
 		return Optional.ofNullable(backup);
 	}
 
@@ -140,132 +130,6 @@ final class SystemPropertiesExtension
 		 * Store entry is for a complete backup object.
 		 */
 		COMPLETE
-	}
-
-	/**
-	 * A sequence of deferred modification applied to a
-	 * {@link Properties} object represented as a single modification.
-	 */
-	private static class DeferredPropertyModification {
-		private static final Object REMOVED = new Object();
-		private final Map<String, Object> changes = new HashMap<>();
-
-		void clearProperty(String key) {
-			changes.put(key, REMOVED);
-		}
-
-		void setProperty(String key, Object value) {
-			changes.put(key, value);
-		}
-
-		void applyTo(Properties properties) {
-			changes.forEach((key, value) -> {
-				// For consistency don't use Properties::setProperty here
-				if (REMOVED.equals(value)) {
-					properties.remove(key);
-				}
-				else {
-					properties.put(key, value);
-				}
-			});
-		}
-
-		/**
-		 * Creates the inverse of calling {@link #applyTo(Properties)} such
-		 * that {@code modification.applyTo(properties); inverse.applyTo(properties);}
-		 * has no observable effect.
-		 */
-		DeferredPropertyModification createInverseApplyTo(Properties properties) {
-			DeferredPropertyModification inverse = new DeferredPropertyModification();
-			changes.keySet().forEach(key -> {
-				// Do not use Properties::getProperty here, since this would
-				// prevent backing up non-string values.
-				Object backup = properties.get(key);
-				if (backup == null) {
-					inverse.clearProperty(key);
-				}
-				else {
-					inverse.setProperty(key, backup);
-				}
-			});
-			return inverse;
-		}
-
-		@Override
-		public String toString() {
-			var builder = new ToStringBuilder(DeferredPropertyModification.class);
-			this.changes.forEach((key, value) -> {
-				if (REMOVED.equals(value)) {
-					builder.append(key, null);
-				}
-				else {
-					builder.append(key, value);
-				}
-			});
-			return builder.toString();
-		}
-
-		static DeferredPropertyModification create(List<ExtensionContext> allContexts) {
-			var modification = new DeferredPropertyModification();
-			// we have to apply the annotations from the outermost to the innermost context.
-			forEachInReverseOrder(allContexts, currentContext -> currentContext.getElement().ifPresent(element -> {
-				var entriesToClear = findEntriesToClear(element);
-				var entriesToSet = findEntriesToSet(element);
-
-				if (entriesToClear.isEmpty() && entriesToSet.isEmpty()) {
-					return;
-				}
-
-				requireNoClearAndSetSameEntries(element, entriesToClear, entriesToSet.keySet());
-				entriesToClear.forEach(modification::clearProperty);
-				entriesToSet.forEach(modification::setProperty);
-			}));
-			return modification;
-		}
-
-		private static Set<String> findEntriesToClear(AnnotatedElement element) {
-			return findRepeatableAnnotations(element, ClearSystemProperty.class).stream() //
-					.map(ClearSystemProperty::key) //
-					// already distinct due to findRepeatableAnnotations
-					.collect(toSet());
-		}
-
-		private static Map<String, String> findEntriesToSet(AnnotatedElement element) {
-			var entries = new HashMap<String, String>();
-			var duplicatePropertyNames = new HashSet<String>();
-
-			findRepeatableAnnotations(element, SetSystemProperty.class) //
-					.forEach(annotation -> {
-						var key = annotation.key();
-						if (entries.put(key, annotation.value()) != null) {
-							duplicatePropertyNames.add(key);
-						}
-					});
-
-			requireUniqueEntries(element, duplicatePropertyNames);
-			return entries;
-		}
-
-		private static void requireNoClearAndSetSameEntries(AnnotatedElement element, Set<String> entriesToClear,
-				Set<String> entriesToSet) {
-			requireUniqueEntries(element, //
-				entriesToClear.stream() //
-						.filter(entriesToSet::contains) //
-						.collect(toSet()));
-		}
-
-		private static void requireUniqueEntries(AnnotatedElement annotatedElement, Set<String> duplicatePropertNames) {
-			if (duplicatePropertNames.isEmpty()) {
-				return;
-			}
-			throw new ExtensionConfigurationException(
-				"SystemPropertyExtension was configured to set/clear %s [%s] more than once by [%s]." //
-						.formatted( //
-							duplicatePropertNames.size() == 1 ? "property" : "properties", //
-							String.join(", ", duplicatePropertNames), //
-							annotatedElement //
-						));
-		}
 	}
 
 	private static List<ExtensionContext> findAllExtensionContexts(ExtensionContext context) {
